Controller & Contents refactors
