+++ packages/gestalt/src/Contents.test.js
@@ -1,14 +1,14 @@
 // @flow strict
 import {
-  getMainDir,
-  getSubDir,
+  getFlyoutDir,
+  getCaretDir,
   calcEdgeShifts,
   baseOffsets,
   adjustOffsets,
   CARET_HEIGHT,
   CARET_WIDTH,
   BORDER_RADIUS,
-} from './Contents.js';
+} from './utils/positioningUtils.js';
 
 const windowSize = {
   height: 900,
@@ -64,16 +64,16 @@ describe('Contents', () => {
   describe('Main Direction chosen correctly', () => {
     it('Chooses the main direction as idealDirection when it fits on screen', () => {
       const triggerRect = centerTriggerRect();
-      idealDirections.forEach((idealDir) => {
-        const mainDir = getMainDir(flyoutSize, idealDir, triggerRect, windowSize);
-        expect(mainDir).toEqual(idealDir);
+      idealDirections.forEach((idealDirection) => {
+        const mainDir = getFlyoutDir({ flyoutSize, idealDirection, triggerRect, windowSize });
+        expect(mainDir).toEqual(idealDirection);
       });
     });
 
     it('Opens down when the trigger is too close to the top of screen', () => {
       const triggerRect = centerTriggerRect({ bottom: 40, top: 0 });
-      idealDirections.forEach((idealDir) => {
-        const mainDir = getMainDir(flyoutSize, idealDir, triggerRect, windowSize);
+      idealDirections.forEach((idealDirection) => {
+        const mainDir = getFlyoutDir({ flyoutSize, idealDirection, triggerRect, windowSize });
         expect(mainDir).toEqual('down');
       });
     });
@@ -83,15 +83,15 @@ describe('Contents', () => {
         bottom: windowSize.height,
         top: windowSize.height - 40,
       });
-      idealDirections.forEach((idealDir) => {
-        const mainDir = getMainDir(flyoutSize, idealDir, triggerRect, windowSize);
+      idealDirections.forEach((idealDirection) => {
+        const mainDir = getFlyoutDir({ flyoutSize, idealDirection, triggerRect, windowSize });
         expect(mainDir).toEqual('up');
       });
     });
 
     it('Chooses the direction in which there is the most space if idealDirection is not given', () => {
       const triggerRect = upperMiddleTriggerRect({ left: 40, right: 80 });
-      const mainDir = getMainDir(flyoutSize, null, triggerRect, windowSize);
+      const mainDir = getFlyoutDir({ flyoutSize, idealDirection: null, triggerRect, windowSize });
       expect(mainDir).toEqual('down');
     });
   });
@@ -99,8 +99,8 @@ describe('Contents', () => {
   describe('Sub direction chosen correctly', () => {
     it('Chooses the middle as sub direction when it fits on the screen', () => {
       const triggerRect = centerTriggerRect();
-      idealDirections.forEach((idealDir) => {
-        const subDir = getSubDir(flyoutSize, idealDir, triggerRect, windowSize);
+      idealDirections.forEach((flyoutDir) => {
+        const subDir = getCaretDir({ flyoutSize, flyoutDir, triggerRect, windowSize });
         expect(subDir).toEqual('middle');
       });
     });
@@ -110,72 +110,114 @@ describe('Contents', () => {
     it('Left opening flyouts', () => {
       const hasCaret = true;
       const triggerRect = centerTriggerRect();
-      const mainDir = 'left';
+      const flyoutDir = 'left';
       const expectedBase = {
         top: windowSize.scrollY + triggerRect.top,
         left: windowSize.scrollX + (triggerRect.left - flyoutSize.width - CARET_HEIGHT),
       };
-      const base = baseOffsets(hasCaret, fixedOffset, flyoutSize, mainDir, triggerRect, windowSize);
+      const base = baseOffsets({
+        hasCaret,
+        relativeOffset: fixedOffset,
+        flyoutSize,
+        flyoutDir,
+        triggerRect,
+        windowSize,
+      });
       expect(base).toEqual(expectedBase);
     });
 
     it('Right opening flyouts', () => {
       const hasCaret = true;
       const triggerRect = centerTriggerRect();
-      const mainDir = 'right';
+      const flyoutDir = 'right';
       const expectedBase = {
         top: windowSize.scrollY + triggerRect.top,
         left: windowSize.scrollX + triggerRect.right + CARET_HEIGHT,
       };
-      const base = baseOffsets(hasCaret, fixedOffset, flyoutSize, mainDir, triggerRect, windowSize);
+      const base = baseOffsets({
+        hasCaret,
+        relativeOffset: fixedOffset,
+        flyoutSize,
+        flyoutDir,
+        triggerRect,
+        windowSize,
+      });
       expect(base).toEqual(expectedBase);
     });
 
     it('Up opening flyouts', () => {
       const hasCaret = true;
       const triggerRect = centerTriggerRect();
-      const mainDir = 'up';
+      const flyoutDir = 'up';
       const expectedBase = {
         top: windowSize.scrollY + (triggerRect.top - flyoutSize.height - CARET_HEIGHT),
         left: windowSize.scrollX + triggerRect.left,
       };
-      const base = baseOffsets(hasCaret, fixedOffset, flyoutSize, mainDir, triggerRect, windowSize);
+      const base = baseOffsets({
+        hasCaret,
+        relativeOffset: fixedOffset,
+        flyoutSize,
+        flyoutDir,
+        triggerRect,
+        windowSize,
+      });
       expect(base).toEqual(expectedBase);
     });
 
     it('Down opening flyouts', () => {
       const hasCaret = true;
       const triggerRect = centerTriggerRect();
-      const mainDir = 'down';
+      const flyoutDir = 'down';
       const expectedBase = {
         top: windowSize.scrollY + triggerRect.bottom + CARET_HEIGHT,
         left: windowSize.scrollX + triggerRect.left,
       };
-      const base = baseOffsets(hasCaret, fixedOffset, flyoutSize, mainDir, triggerRect, windowSize);
+      const base = baseOffsets({
+        hasCaret,
+        relativeOffset: fixedOffset,
+        flyoutSize,
+        flyoutDir,
+        triggerRect,
+        windowSize,
+      });
       expect(base).toEqual(expectedBase);
     });
 
     it('Left opening flyouts without caret', () => {
       const hasCaret = false;
       const triggerRect = centerTriggerRect();
-      const mainDir = 'left';
+      const flyoutDir = 'left';
       const expectedBase = {
         top: windowSize.scrollY + triggerRect.top,
         left: windowSize.scrollX + (triggerRect.left - flyoutSize.width - 8),
       };
-      const base = baseOffsets(hasCaret, fixedOffset, flyoutSize, mainDir, triggerRect, windowSize);
+      const base = baseOffsets({
+        hasCaret,
+        relativeOffset: fixedOffset,
+        flyoutSize,
+        flyoutDir,
+        triggerRect,
+        windowSize,
+      });
       expect(base).toEqual(expectedBase);
     });
 
     it('Down opening flyouts without caret', () => {
       const hasCaret = false;
       const triggerRect = centerTriggerRect();
-      const mainDir = 'down';
+      const flyoutDir = 'down';
       const expectedBase = {
         top: windowSize.scrollY + triggerRect.bottom + 8,
         left: windowSize.scrollX + triggerRect.left,
       };
-      const base = baseOffsets(hasCaret, fixedOffset, flyoutSize, mainDir, triggerRect, windowSize);
+      const base = baseOffsets({
+        hasCaret,
+        relativeOffset: fixedOffset,
+        flyoutSize,
+        flyoutDir,
+        triggerRect,
+        windowSize,
+      });
       expect(base).toEqual(expectedBase);
     });
   });
@@ -183,8 +225,8 @@ describe('Contents', () => {
   describe('Adjusted offsets chosen correctly', () => {
     it('Left-up opening flyouts with caret shifted left past rounded corners', () => {
       const triggerRect = centerTriggerRect();
-      const mainDir = 'left';
-      const subDir = 'up';
+      const flyoutDir = 'left';
+      const caretDir = 'up';
       const base = {
         top: 100,
         left: 100,
@@ -209,14 +251,15 @@ describe('Contents', () => {
         bottom: null,
         left: null,
       };
-      const { flyoutOffset, caretOffset } = adjustOffsets(
+      const { flyoutOffset, caretOffset } = adjustOffsets({
         base,
         edgeShift,
         flyoutSize,
-        mainDir,
-        subDir,
+        flyoutDir,
+        caretDir,
         triggerRect,
-      );
+        isScrollableBox: false,
+      });
       expect(flyoutOffset).toEqual(expectedFlyoutOffset);
       expect(caretOffset).toEqual(expectedCaretOffset);
     });
@@ -224,8 +267,8 @@ describe('Contents', () => {
 
   it('Right-up opening flyouts with caret shifted right past rounded corners', () => {
     const triggerRect = centerTriggerRect();
-    const mainDir = 'right';
-    const subDir = 'up';
+    const flyoutDir = 'right';
+    const caretDir = 'up';
     const base = {
       top: 100,
       left: 100,
@@ -250,14 +293,15 @@ describe('Contents', () => {
       bottom: null,
       left: -CARET_HEIGHT,
     };
-    const { flyoutOffset, caretOffset } = adjustOffsets(
+    const { flyoutOffset, caretOffset } = adjustOffsets({
       base,
       edgeShift,
       flyoutSize,
-      mainDir,
-      subDir,
+      flyoutDir,
+      caretDir,
       triggerRect,
-    );
+      isScrollableBox: false,
+    });
     expect(flyoutOffset).toEqual(expectedFlyoutOffset);
     expect(caretOffset).toEqual(expectedCaretOffset);
   });
@@ -265,8 +309,7 @@ describe('Contents', () => {
   describe('Edge shifts calculated correctly', () => {
     it('Keeps Container on screen when trigger is on the edge', () => {
       const triggerRect = upperLeftTriggerRect();
-      const subDir = 'up';
-      const { flyout, caret } = calcEdgeShifts(subDir, triggerRect, windowSize);
+      const { flyout, caret } = calcEdgeShifts({ triggerRect, windowSize, isScrollableBox: false });
       expect(triggerRect.bottom - flyout.y).toBeGreaterThan(0);
       expect(flyout.x).toBeLessThan(BORDER_RADIUS);
       expect(caret.x).toEqual(caret.y);
